Made a bunch of patterns for subpixels
Added Color Correction, this divides the color output by a color.

So if you're upping the red channel in post processing, set the color to red to get the screen back to normal colors.

Added Alpha channel support for the subpixels to control which parts of a sub pixel light up in what order.

The only pattern that has this at the moment is the modern LCD one, as it's the only one that needs it.

It's just a grayscale alpha channel, 0 to 1, 1 will light up first, while 0.1 will light up last.